Separate Standard from Noaa.
- I didn't make the Noaa interface and impl a jar, so it might be
confusing if we have our "Standard" Implementation in the vendor
package.

Just want to drive the point home that vendor files are something we
don't touch.
